+++ elasticsearch/src/main/java/com/tcdi/zombodb/query_parser/ZomboDBQueryRewriter.java
@@ -16,10 +16,6 @@
 package com.tcdi.zombodb.query_parser;
 
 import org.elasticsearch.client.Client;
-import org.elasticsearch.index.query.BoolQueryBuilder;
-import org.elasticsearch.index.query.QueryBuilder;
-
-import static org.elasticsearch.index.query.QueryBuilders.boolQuery;
 
 /**
  * ZomboDB's stock {@link QueryRewriter} that resolves joins during construction
@@ -31,20 +27,8 @@ public class ZomboDBQueryRewriter extends QueryRewriter {
     }
 
     @Override
-    protected void performCustomOptimizations() {
+    protected void performOptimizations(Client client) {
+        super.performOptimizations(client);
         new ExpansionOptimizer(this, tree, metadataManager, client, searchPreference, doFullFieldDataLookup).optimize();
-    }
-
-    @Override
-    protected QueryBuilder build(ASTExpansion node) {
-        QueryBuilder expansionBuilder =  build(node.getQuery());
-        QueryParserNode filterQuery = node.getFilterQuery();
-        if (filterQuery != null) {
-            BoolQueryBuilder bqb = boolQuery();
-            bqb.must(applyExclusion(build(node.getQuery()), node.getIndexLink().getIndexName()));
-            bqb.must(build(filterQuery));
-            expansionBuilder = bqb;
-        }
-        return expansionBuilder;
     }
 }
