+++ elasticsearch/src/test/java/com/tcdi/zombodb/test/ZomboDBTestCase.java
@@ -16,7 +16,6 @@
 package com.tcdi.zombodb.test;
 
 import com.tcdi.zombodb.highlight.AnalyzedField;
-import com.tcdi.zombodb.query_parser.JUnitQueryRewriter;
 import com.tcdi.zombodb.query_parser.QueryRewriter;
 import org.elasticsearch.action.admin.indices.create.CreateIndexRequestBuilder;
 import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
@@ -124,7 +123,7 @@ public abstract class ZomboDBTestCase {
 
 
     protected QueryRewriter qr(String query) {
-        return new JUnitQueryRewriter(client(), DEFAULT_INDEX_NAME, null, query, true, false);
+        return new QueryRewriter(client(), DEFAULT_INDEX_NAME, query, null, true, false) { /* anonymous implementation */ };
     }
 
     protected void assertJson(String query, String expectedJson) throws Exception {
